Split deployment_process out into separate files and got it working
